toolbox: convert the toolbox into a template tag
This is a *huge* improvement to how the code is organized right now.
Specifically this gives us complete control over how stuff is loaded from the
sidebar and makes more evident which are the paths taken by the code to
actually do stuff. A concrete example is how loading of links happen.

amancay_interface.js has been disabled to implement/move its functionality to
other places.

git-svn-id: svn+ssh://svn.debian.org/svn/bts-webui/bts_webui@145 19058424-5d2e-0410-ba3e-95a0a9172157